feat(phase4): math/rand test suite validates in C# (package #15)
math/rand's Go test suite validates through the -tests pipeline: 43 tests match
`go test -json` with 0 skipped on either side and ZERO disclosed divergences (22
Example/Benchmark declarations excluded as the standard Phase-4D deferral). No
numeric or semantic mismatch anywhere.

Notably this includes TestDefaultRace and all 6 of its parallel subtests, which
re-execute the test binary as a child process. That needed real os/exec child-process
creation on Windows -- argv + environment passing, stdout/stderr capture through
CombinedOutput, wait, and exit-code surfacing -- which landed in this same stack. It
was briefly a candidate for a "capability-gap deferral"; fixing the capability made
the deferral unnecessary, which is the outcome that policy's guardrail was written to
force.

Four converter/shim fixes in this stack unblocked the suite's compile, each general
rather than rand-specific: the compile-only testing PB/RunParallel surface; anchoring a
NESTED package's production GoImplement record to the production metadata file (every
math/*, text/*, net/*, encoding/*, container/*, hash/*, crypto/* external test that
casts a production type to an interface was broken -- top-level packages only worked by
accident); scanning converted test sources for package-qualifier usings when resolving
test project references; and narrowing a constant expression whose folded subexpression
widened it past the target type (which also repaired latent ulong/long defects in nine
stdlib packages and makes rand's own Int31n compute in uint32 exactly as Go does).

Committed per the validated-package policy: math/rand production C# refreshed to the
current converter, its test artifacts (rand_test/regress_test/auto_test/default_test/
race_test/example_test/export_test, package_test_info, package_info_test,
go2cs_test_host, math.rand.tests.csproj) for reviewability, and the production csproj's
test-artifact exclusion. The pipeline's regenerated inputs stay git-ignored.

Verified on the full merged stack: CNR byte-identical across 434 behavioral projects,
and the complete 302-package corpus builds with ZERO errors.

+++ src/go-src-converted/math/rand/rand.cs
@@ -156,7 +156,7 @@ public static void Seed(this ж<Rand> Ꮡr, int64 seed) {
         // n is power of two, can mask
         return (int32)(r.Int31() & (n - 1));
     }
-    var max = (int32)((2147483648L) - 1 - (((uint32)1 << (int)(31))) % (uint32)n);
+    var max = (int32)((uint32)((2147483648L) - 1) - (((uint32)1 << (int)(31))) % (uint32)n);
     var v = r.Int31();
     while (v > max) {
         v = r.Int31();
@@ -194,7 +194,7 @@ public static void Seed(this ж<Rand> Ꮡr, int64 seed) {
     if (n <= 0) {
         throw panic("invalid argument to Intn");
     }
-    if (n <= 2147483648L - 1) {
+    if (n <= (nint)(2147483648L - 1)) {
         return (nint)r.Int31n((int32)n);
     }
     return (nint)r.Int63n((int64)n);
